+++ Editor/Util.cs
@@ -1135,6 +1135,29 @@ namespace Reallusion.Import
             return false;
         }
 
+
+
+
+
+
+        public static GameObject EditPrefabContents(GameObject prefabAsset)
+        {
+            GameObject prefabRoot;
+            string currentPrefabAssetPath = AssetDatabase.GetAssetPath(prefabAsset);
+            prefabRoot = PrefabUtility.LoadPrefabContents(currentPrefabAssetPath);
+            return prefabRoot;
+        }
+
+        public static void SaveAndUnloadPrefabContents(GameObject prefabAsset, GameObject prefabContents)
+        {
+            string currentPrefabAssetPath = AssetDatabase.GetAssetPath(prefabAsset);
+            PrefabUtility.SaveAsPrefabAsset(prefabContents, currentPrefabAssetPath, out bool success);
+            PrefabUtility.UnloadPrefabContents(prefabContents);
+        }
+
+
+
+
         private static Editor MakeEditor(string guid)
         {
             Object o = AssetDatabase.LoadAssetAtPath(AssetDatabase.GUIDToAssetPath(guid), typeof(Object));